Roting tables updating optimization: remove saving GatewayNotificationMessage into storage from ConfirmationRequiredMessagesHandler

+++ src/core/network/communicator/internal/queue/ConfirmationRequiredMessagesHandler.cpp
@@ -98,10 +98,12 @@ void ConfirmationRequiredMessagesHandler::tryProcessConfirmation(
                    << queueKey.first;
         }
 
-        auto ioTransaction = mCommunicatorStorageHandler->beginTransaction();
-        ioTransaction->communicatorMessagesQueueHandler()->deleteRecord(
-            queueKey.second,
-            confirmationMessage->transactionUUID());
+        if (confirmationMessage->typeID() != Message::GatewayNotification) {
+            auto ioTransaction = mCommunicatorStorageHandler->beginTransaction();
+            ioTransaction->communicatorMessagesQueueHandler()->deleteRecord(
+                queueKey.second,
+                confirmationMessage->transactionUUID());
+        }
 
 #ifdef DEBUG_LOG_NETWORK_COMMUNICATOR
         debug() << "Confirmation for message with transaction UUID " << confirmationMessage->transactionUUID() << " received. "
